joint path generator

+++ src/moveEngine/AbstractPathGenerator.h
@@ -9,9 +9,9 @@ public:
     // AbstractPathGenerator();
     virtual ~AbstractPathGenerator(){};
 
-    virtual void setCurrentPoint(int* currentPoint) = 0;
-    virtual void setTargetPoint(int* targetPoint) = 0;
-    virtual void setLimits(int acceleration, int speedLimit) = 0;
+    virtual void setCurrentPoint(float* currentPoint) = 0;
+    virtual void setTargetPoint(float* targetPoint) = 0;
+    virtual void setLimits(float acceleration, float speedLimit) = 0;
     virtual void computeTrajectory(Trajectory *trajectory) = 0;
 };
